CTCTRADERS-143 Split navigator spec into two test files.
- Two spec files, one for NormalMode and one for CheckMode.
